address review: use Class_ node type, skip rename on existing method name collision
