+++ rules-tests/CodeQuality/Rector/Class_/PreferTestsWithCamelCaseRector/config/configured_rule.php
@@ -3,7 +3,7 @@
 declare(strict_types=1);
 
 use Rector\Config\RectorConfig;
-use Rector\PHPUnit\CodeQuality\Rector\ClassMethod\PreferTestsWithCamelCaseRector;
+use Rector\PHPUnit\CodeQuality\Rector\Class_\PreferTestsWithCamelCaseRector;
 
 return static function (RectorConfig $rectorConfig): void {
     $rectorConfig->rule(PreferTestsWithCamelCaseRector::class);
