+++ rules-tests/CodeQuality/Rector/Class_/PreferTestsWithSnakeCaseRector/Fixture/skip_existing_name.php.inc
@@ -1,0 +1,16 @@
+<?php
+
+namespace Rector\PHPUnit\Tests\CodeQuality\Rector\Class_\PreferTestsWithSnakeCaseRector\Fixture;
+
+use PHPUnit\Framework\TestCase;
+
+final class SkipExistingNameTest extends TestCase
+{
+    public function testMe()
+    {
+    }
+
+    public function test_me()
+    {
+    }
+}
