+++ rules-tests/CodeQuality/Rector/Class_/PreferTestsWithSnakeCaseRector/config/configured_rule.php
@@ -3,7 +3,7 @@
 declare(strict_types=1);
 
 use Rector\Config\RectorConfig;
-use Rector\PHPUnit\CodeQuality\Rector\ClassMethod\PreferTestsWithSnakeCaseRector;
+use Rector\PHPUnit\CodeQuality\Rector\Class_\PreferTestsWithSnakeCaseRector;
 
 return static function (RectorConfig $rectorConfig): void {
     $rectorConfig->rule(PreferTestsWithSnakeCaseRector::class);
